# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "深地智学 API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/external-resources": {
      "get": {
        "tags": [
          "external-resources"
        ],
        "summary": "Get Resources",
        "operationId": "get_resources_api_v1_external_resources_get",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ResourceCategory"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 255
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider"
            }
          },
          {
            "name": "course_level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 32
                },
                {
                  "type": "null"
                }
              ],
              "title": "Course Level"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExternalResource"
                  },
                  "title": "Response Get Resources Api V1 External Resources Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/geo-data/datasets": {
      "get": {
        "tags": [
          "geo-data"
        ],
        "summary": "Get Geo Datasets",
        "operationId": "get_geo_datasets_api_v1_geo_data_datasets_get",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/GeoDatasetCategory"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GeoDataset"
                  },
                  "title": "Response Get Geo Datasets Api V1 Geo Data Datasets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/geo-data/datasets/{dataset_id}": {
      "get": {
        "tags": [
          "geo-data"
        ],
        "summary": "Get Geo Dataset Detail",
        "operationId": "get_geo_dataset_detail_api_v1_geo_data_datasets__dataset_id__get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeoDatasetDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/search": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Search Knowledge",
        "operationId": "search_knowledge_api_v1_knowledge_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeSearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/documents": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Get Documents",
        "operationId": "get_documents_api_v1_knowledge_documents_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DocumentStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Document Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KnowledgeDocument"
                  },
                  "title": "Response Get Documents Api V1 Knowledge Documents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Upload Document",
        "operationId": "upload_document_api_v1_knowledge_documents_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_knowledge_documents_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeDocument"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/documents/{document_id}/transitions": {
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Update Document Status",
        "operationId": "update_document_status_api_v1_knowledge_documents__document_id__transitions_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentTransitionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeDocument"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/external-resources": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Add External Resource",
        "operationId": "add_external_resource_api_v1_admin_external_resources_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalResourceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalResource"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_document_api_v1_knowledge_documents_post": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "document_type": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Document Type"
          },
          "course_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Name"
          },
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "title",
          "document_type",
          "file"
        ],
        "title": "Body_upload_document_api_v1_knowledge_documents_post"
      },
...
      "Citation": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "course_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Name"
          },
          "source_locator": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Locator"
          },
          "excerpt": {
            "type": "string",
            "title": "Excerpt"
          },
          "score": {
            "type": "number",
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "title",
          "excerpt",
          "score"
        ],
        "title": "Citation"
      },
...
      "DocumentStatus": {
        "type": "string",
        "enum": [
          "uploaded",
          "parsed",
          "reviewed",
          "indexed",
          "published",
          "rejected"
        ],
        "title": "DocumentStatus"
      },
      "DocumentTransitionRequest": {
        "properties": {
          "target_status": {
            "$ref": "#/components/schemas/DocumentStatus"
          }
        },
        "type": "object",
        "required": [
          "target_status"
        ],
        "title": "DocumentTransitionRequest"
      },
      "ExternalResource": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "category": {
            "$ref": "#/components/schemas/ResourceCategory"
          },
          "course_level": {
            "type": "string",
            "title": "Course Level"
          },
          "audience": {
            "type": "string",
            "title": "Audience"
          },
          "credits": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credits"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "中文"
          },
          "status": {
            "$ref": "#/components/schemas/ResourceStatus",
            "default": "active"
          },
          "url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "cover_asset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Asset"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "provider",
          "category",
          "course_level",
          "audience",
          "description",
          "sort_order"
        ],
        "title": "ExternalResource"
      },
      "ExternalResourceCreate": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "category": {
            "$ref": "#/components/schemas/ResourceCategory"
          },
          "course_level": {
            "type": "string",
            "title": "Course Level"
          },
          "audience": {
            "type": "string",
            "title": "Audience"
          },
          "credits": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credits"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "中文"
          },
          "status": {
            "$ref": "#/components/schemas/ResourceStatus",
            "default": "active"
          },
          "url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "cover_asset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Asset"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "provider",
          "category",
          "course_level",
          "audience",
          "description"
        ],
        "title": "ExternalResourceCreate"
      },
      "GeoDataset": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "category": {
            "$ref": "#/components/schemas/GeoDatasetCategory"
          },
          "region": {
            "type": "string",
            "title": "Region"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "data_format": {
            "type": "string",
            "title": "Data Format"
          },
          "resolution": {
            "type": "string",
            "title": "Resolution"
          },
          "size_label": {
            "type": "string",
            "title": "Size Label"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "license": {
            "type": "string",
            "title": "License"
          },
          "access_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Url"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order"
          },
          "feature_count": {
            "type": "integer",
            "title": "Feature Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "category",
          "region",
          "description",
          "data_format",
          "resolution",
          "size_label",
          "source",
          "license",
          "sort_order"
        ],
        "title": "GeoDataset"
      },
      "GeoDatasetCategory": {
        "type": "string",
        "enum": [
          "geology",
          "borehole",
          "logging",
          "seismic",
          "geochemistry",
          "remote_sensing",
          "terrain",
          "potential_field"
        ],
        "title": "GeoDatasetCategory"
      },
      "GeoDatasetDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "category": {
            "$ref": "#/components/schemas/GeoDatasetCategory"
          },
          "region": {
            "type": "string",
            "title": "Region"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "data_format": {
            "type": "string",
            "title": "Data Format"
          },
          "resolution": {
            "type": "string",
            "title": "Resolution"
          },
          "size_label": {
            "type": "string",
            "title": "Size Label"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "license": {
            "type": "string",
            "title": "License"
          },
          "access_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Url"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order"
          },
          "feature_count": {
            "type": "integer",
            "title": "Feature Count",
            "default": 0
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/GeoDatasetFeature"
            },
            "type": "array",
            "title": "Features",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "category",
          "region",
          "description",
          "data_format",
          "resolution",
          "size_label",
          "source",
          "license",
          "sort_order"
        ],
        "title": "GeoDatasetDetail"
      },
      "GeoDatasetFeature": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "lon": {
            "type": "number",
            "maximum": 136.0,
            "minimum": 73.0,
            "title": "Lon"
          },
          "lat": {
            "type": "number",
            "maximum": 54.0,
            "minimum": 3.0,
            "title": "Lat"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "name",
          "lon",
          "lat"
        ],
        "title": "GeoDatasetFeature",
        "description": "数据集内公开的示例点位（教学示意，非工业成果）。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "KnowledgeDocument": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "document_type": {
            "type": "string",
            "title": "Document Type"
          },
          "course_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Name"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "checksum": {
            "type": "string",
            "title": "Checksum"
          },
          "byte_size": {
            "type": "integer",
            "title": "Byte Size"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "parsing_attempts": {
            "type": "integer",
            "title": "Parsing Attempts"
          },
          "processing_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Error"
          },
          "parsed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parsed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "document_type",
          "original_filename",
          "content_type",
          "checksum",
          "byte_size",
          "status",
          "version",
          "created_by",
          "parsing_attempts",
          "created_at",
          "updated_at"
        ],
        "title": "KnowledgeDocument"
      },
...
      "KnowledgeSearchResponse": {
        "properties": {
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "title": "Citations"
          }
        },
        "type": "object",
        "required": [
          "citations"
        ],
        "title": "KnowledgeSearchResponse"
      },
...
      "ResourceCategory": {
        "type": "string",
        "enum": [
          "courses",
          "knowledge",
          "practice"
        ],
        "title": "ResourceCategory"
      },
      "ResourceStatus": {
        "type": "string",
        "enum": [
          "active",
          "pending"
        ],
        "title": "ResourceStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}